% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.sbygroup.R
\name{mice.impute.sbygroup}
\alias{mice.impute.sbygroup}
\title{Imputation by group, modified from https://github.com/alexanderrobitzsch/miceadds}
\usage{
mice.impute.sbygroup(y, ry, x, wy = NULL, group, imputationFunction, ...)
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes missing data by group,
}
\details{
This is the same as thr original mice.impute.bygroup function,
except that the spolr.fit attribute is attached to imputations.
The spolr.fit attribute is a list with the length of the number
of groups, whereby each entry of the list is a spolr.fit object for
a group
}
\seealso{
Other univariate imputation functions: \code{\link{mice.impute.sbinomial}},
  \code{\link{mice.impute.sgauss}},
  \code{\link{mice.impute.slogreg}},
  \code{\link{mice.impute.spolr}}
}
\author{
Alexander Robitzsch with minor modification by Guido Biele
}
\concept{univariate imputation functions}
\keyword{datagen}
