% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.sgauss.R
\name{mice.impute.sgauss}
\alias{mice.impute.sgauss}
\title{Imputation of ordered data by bayesian ordered logistic regression}
\usage{
mice.impute.sgauss(y, ry, x, wy = NULL, ...)
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes missing data in a normally distributed variable
}
\details{
The function \code{mice.impute.gauss()} imputes for normal
variables by using the gaussian regression model. The model is fit in Stan and
uses weakly informative (shrinkage) priors to regularize.
}
\seealso{
Other univariate imputation functions: \code{\link{mice.impute.sbinomial}},
  \code{\link{mice.impute.sbygroup}},
  \code{\link{mice.impute.slogreg}},
  \code{\link{mice.impute.spolr}}
}
\author{
Guido Biele
}
\concept{univariate imputation functions}
\keyword{datagen}
