% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.snegbin.R
\name{mice.impute.snegbin}
\alias{mice.impute.snegbin}
\title{Imputation using bayesian regularized regression}
\usage{
mice.impute.snegbin(y, ry, x, wy = NULL, ...)
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes missing data in a count variable using negative binomial regression
}
\details{
The function \code{mice.impute.snegbin()} imputes for binomial response
variables by using the negative binomial regression with a logit link function.
The model is fit in Stan and uses weakly informative (shrinkage) priors to
regularize regression coefficients.
}
\seealso{
Other univariate imputation functions: \code{\link{mice.impute.sbetabin}},
  \code{\link{mice.impute.sbinomial}},
  \code{\link{mice.impute.sbygroup}},
  \code{\link{mice.impute.sgauss}},
  \code{\link{mice.impute.slogreg}},
  \code{\link{mice.impute.spolr}}
}
\author{
Guido Biele
}
\concept{univariate imputation functions}
\keyword{datagen}
