% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.spolr.R
\name{mice.impute.spolr}
\alias{mice.impute.spolr}
\title{Imputation of ordered data by bayesian ordered logistic regression}
\usage{
mice.impute.spolr(y, ry, x, wy = NULL, ...)
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes missing data in a categorical variable using polytomous regression
}
\details{
The function \code{mice.impute.spolr()} imputes for ordered categorical response
variables by using the proportional odds logistic regression (polr) model. The
model is also known as the cumulative link model. The model is fit in Stan and
uses weakly informative (shrinkage) priors to regularize
}
\seealso{
Other univariate imputation functions: \code{\link{mice.impute.sbinomial}},
  \code{\link{mice.impute.sbygroup}},
  \code{\link{mice.impute.sgauss}},
  \code{\link{mice.impute.slogreg}}
}
\author{
Guido Biele
}
\concept{univariate imputation functions}
\keyword{datagen}
