% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbetabin.R
\name{predict.sbetabin}
\alias{predict.sbetabin}
\title{Prediction for snegbin}
\usage{
\method{predict}{sbetabin}(object, newdata, type = c("response",
  "linear"), method = "logistic")
}
\arguments{
\item{object}{Object of class `spolr` (returned from function `spolr`).}

\item{newdata}{data.frame with new data (optional).}

\item{type}{kind of predictions ( `response` or `linear`).}
}
\value{
predicted reponses
}
\description{
Prediction for snegbin
}
