% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbinomial.R
\name{predict.sbinomial}
\alias{predict.sbinomial}
\title{Prediction for sbinomial}
\usage{
\method{predict}{sbinomial}(object, newdata, type = c("response",
  "linear"), method = "logistic")
}
\arguments{
\item{object}{Object of class `spolr` (returned from function `spolr`).}

\item{newdata}{data.frame with new data (optional).}

\item{type}{kind of predictions (probabilities, `probs` - or highest probability response, `class`).}
}
\value{
predicted reponses
}
\description{
Prediction for sbinomial
}
