% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgauss.R
\name{predict.sgauss}
\alias{predict.sgauss}
\title{Prediction for sgauss}
\usage{
\method{predict}{sgauss}(object, newdata)
}
\arguments{
\item{object}{Object of class `sgauss` (returned from function `sgauss`).}

\item{newdata}{data.frame with new data (optional).}
}
\value{
predicted reponses
}
\description{
Prediction for sgauss
}
