% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slogreg.R
\name{predict.slogreg}
\alias{predict.slogreg}
\title{Prediction for slogreg}
\usage{
\method{predict}{slogreg}(object, newdata, type = c("probs", "class"))
}
\arguments{
\item{object}{Object of class `spolr` (returned from function `spolr`).}

\item{newdata}{data.frame with new data (optional).}

\item{type}{kind of predictions (probabilities, `probs` - or highest probability response, `class`).}
}
\value{
predicted reponses
}
\description{
Prediction for slogreg
}
