% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spolr.R
\name{predict.spolr}
\alias{predict.spolr}
\title{Prediction for sporl}
\usage{
\method{predict}{spolr}(object, newdata, type = c("probs", "class"),
  method = "logistic")
}
\arguments{
\item{object}{Object of class `spolr` (returned from function `spolr`).}

\item{newdata}{data.frame with new data (optional).}

\item{type}{kind of predictions (probabilities, `probs` - or highest probability response, `class`).}
}
\value{
predicted reponses
}
\description{
Prediction for sporl
}
