% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spolr-package.R
\docType{package}
\name{spolr-package}
\alias{spolr-package}
\alias{spolr}
\title{The 'spolr' package.}
\description{
Bayesian logistic and ordinal regression for mice (includes mice.impute...)
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
